requirements.txt
flask-sqlalchemy
mysqlclient
configparser
flask-mysql
mysql